+++ src/Xteq5OutputGenerator_Consumer/Xteq5OutputGeneratorConsumerProgram.cs
@@ -20,8 +20,8 @@ namespace Xteq5OutputGenerator_Consumer
 
             
             //basepath = @"C:\dev\git\xteq5\src\ScriptsForTesting\Test5";
-            basepath = @"C:\dev\git\xteq5\src\ScriptsForTesting\Test1";
-            //basepath = @"C:\dev\git\xteq5\scripts";
+            //basepath = @"C:\dev\git\xteq5\src\ScriptsForTesting\Test1";
+            basepath = @"C:\dev\git\xteq5\scripts";
 
 
             Task<Report> task = RunEngineAsync(runner, basepath);
@@ -46,7 +46,7 @@ namespace Xteq5OutputGenerator_Consumer
             File.WriteAllText(filename, result);
             */
 
-            string filename = OutputGenerator.GenerateReportOutputFile(report, OutputFormatEnum.XML, "");
+            string filename = OutputGenerator.GenerateReportOutputFile(report, OutputFormatEnum.HTML, "");
             //string filename = OutputGenerator.GenerateReportOutputFile(report, OutputFormatEnum.HTML, "");
             Process.Start(filename);